↔️ Handle dropped and replaced transactions (#355)
* Handle dropped and replaced transactions

Previously if a user speeded up a transaction the old transaction would report as failed and the new one would not exist in useDapp state. 

This change add handling around failed transaction to check if it was replaced and repriced and should therefore continue on as a new pending transaction.