➕ Add stricter contract type check on the useLogs hook (#829)
* Add stricter contract type check on the useLogs hook

Using generics in the TypedFilter interface add some nice type checks to the filter parameter of useLogs, but it doesn't help the user when creating the parameter, unless they go out of their way to instantiate he object with the TypedFilter type specified. This way, the check is provided directly by the function. It can still be bypassed, if one wishes to, by calling it this way: useLogs<Contract>(...)

* Update PR name

* Create add-sctriter-check-useLogs.md

Co-authored-by: Mateusz Janduła <71319308+mj426382@users.noreply.github.com>